Refresh patches.arch/powerpc-pseries-fix-fadump-endianess-backport-sles12-gmc.patch
The patch uselessly changes type of size variable causing a warning.

suse-commit: 0ea72fa4d24061f40af7405344687ea6e4df9b92

+++ arch/powerpc/kernel/fadump.c
@@ -57,7 +57,7 @@ int __init early_init_dt_scan_fw_dump(unsigned long node,
 {
 	const __be32 *sections;
 	int i, num_sections;
-	int size;
+	unsigned long size;
 	const __be32 *token;
 
 	if (depth != 1 || strcmp(uname, "rtas") != 0)
